requirements.txt
Django==4.1
django-ckeditor==6.5.1
Pillow==9.4.0
